requirements.txt
torch
torchvision
transformers
diffusers
accelerate
bitsandbytes
pyyaml
Pillow
tqdm
